codigo/requirements.txt
kagglehub
pandas
opencv-python
numpy
matplotlib
tensorflow
scikit-learn
faiss-cpu
gradio
gdown
seaborn
ultralytics